Move getSelectors logic into separate module
This allows it to be reused by utilities
